Run Phospho Adjustment

runPhosphoAdjustment performs phospho adjustment on a MultiAssayExperiment object to normalize the phosphoproteome data.

Usage

runPhosphoAdjustment(
  maeData,
  normalization = FALSE,
  minOverlap = 3,
  completeness = 0,
  ncore = 1
)

Arguments

maeData

A MultiAssayExperiment object containing phosphoproteome and full proteome data.

normalization

A logical value indicating whether to perform normalization. Default is FALSE.

minOverlap

A numeric value specifying the minimum number of overlapping peptides required between samples. Default is 3.

completeness

A numeric value indicating the required completeness of data for features to be included. Default is 0.

ncore

A numeric value specifying the number of cores to use for parallel processing. Default is 1.

Value

A MultiAssayExperiment object with adjusted phosphoproteome data.

Details

The function performs the following steps:

Defines an optimization function to minimize the sum of squared differences between pairs of samples.

Calculates the ratio matrix of phosphoproteome to full proteome data.

Subsets features based on completeness criteria.

Performs a sanity check to identify and exclude problematic samples.

Sets initial values for the adjustment factor based on column medians.

Estimates the adjustment factor using parallel optimization.

Adjusts the phosphoproteome measurements using the estimated adjustment factor.